--- app/src/main/java/com/example/sharingbookshelf/Activities/TakingPhotoActivity.java
@@ -1,134 +1,0 @@
-package com.example.sharingbookshelf.Activities;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageView;
-
-import com.example.sharingbookshelf.BookApiRetrofitClient;
-import com.example.sharingbookshelf.HttpRequest.RetrofitServiceApi;
-import com.example.sharingbookshelf.Models.BookApiResponse;
-import com.example.sharingbookshelf.R;
-import com.example.sharingbookshelf.RetrofitClient;
-
-import org.json.JSONArray;
-
-import java.lang.reflect.Array;
-import java.util.ArrayList;
-import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class TakingPhotoActivity extends AppCompatActivity {
-
-    private static final int BARCODE_ACTIVITY = 10000;
-    private static final int ADDSELF_ACTIVITY = 10001;
-    private static final int BOOKPOPUP_ACTIVITY = 10002;
-    private RetrofitServiceApi retrofitServiceApi;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_taking_photo);
-
-        Button btn_barcode = findViewById(R.id.btn_barcode);
-        Button btn_selfAddBook = findViewById(R.id.btn_selfAddBook);
-
-        //바코드로 추가
-        btn_barcode.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(TakingPhotoActivity.this, BarcodeActivity.class);
-                startActivityForResult(intent, BARCODE_ACTIVITY);
-            }
-        });
-
-        //직접추가
-        btn_selfAddBook.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(TakingPhotoActivity.this, SelfAddBookPopupActivity.class);
-                startActivityForResult(intent, ADDSELF_ACTIVITY);
-
-            }
-        });
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
-        super.onActivityResult(requestCode, resultCode, intent);
-        if (requestCode == BARCODE_ACTIVITY) { //바코드 인식 결과
-            if (resultCode == RESULT_OK) {
-                String data = intent.getExtras().getString("ISBN");
-                if (data != null) {
-                    //et_address.setText(data);
-                    Log.d(MainActivity.MAIN_TAG, data);
-                    callBookResponse(data);
-                }
-            }
-        }
-        if (requestCode == ADDSELF_ACTIVITY) { //직접 추가 결과
-            if (resultCode == RESULT_OK) {
-                String data = intent.getExtras().getString("ISBN");
-                if (data != null) {
-                    Log.d(MainActivity.MAIN_TAG, data);
-                    callBookResponse(data);
-                }
-            }
-        }
-    }
-
-    /* Kakao Book search API 통신 */
-    private void callBookResponse(String ISBN) {
-        retrofitServiceApi = BookApiRetrofitClient.createService(RetrofitServiceApi.class);
-        Call<BookApiResponse> call = retrofitServiceApi.setBookApiResponse(ISBN, "isbn");
-        call.enqueue(new Callback<BookApiResponse>() {
-            @Override
-            public void onResponse(Call<BookApiResponse> call, Response<BookApiResponse> response) {
-                BookApiResponse result = response.body();
-                Log.d(MainActivity.MAIN_TAG, "책 api 통신 성공");
-                if (result != null) {
-                    /*for (int i = 0; i < result.documents.size(); i++) {
-                        Log.d(MainActivity.MAIN_TAG,"title : " + result.documents.get(i).getTitle());
-                    }*/
-                    getBookDetails(result);
-                }
-            }
-
-            @Override
-            public void onFailure(Call<BookApiResponse> call, Throwable t) {
-                Log.e(MainActivity.MAIN_TAG, "책 api 통신 실패", t);
-            }
-        });
-    }
-
-    private void getBookDetails(BookApiResponse books) {
-        ArrayList<BookApiResponse.Document> documentList = books.documents;
-        BookApiResponse.Meta meta = books.metas;
-        //BookApiResponse.Document book = (BookApiResponse.Document) books.documents.get(0);
-        Intent intent = new Intent(TakingPhotoActivity.this, BookDetailsPopupActivity.class);
-        intent.putExtra("documentList", documentList);
-        intent.putExtra("meta", meta);
-        startActivityForResult(intent, BOOKPOPUP_ACTIVITY);
-        /*Bundle extra = new Bundle();
-        Intent intent = getIntent();
-        extra.putString("ISBN", ISBN);
-        intent.putExtras(extra);
-        setResult(RESULT_OK, intent);
-        finish();*/
-    }
-
-
-}
-
-
-
-
-
-
