+++ app/src/main/res/layout/activity_home.xml
@@ -51,7 +51,6 @@
                 android:text="nickname"
                 android:textSize="30dp"
                 android:textStyle="bold"
-                android:background="#F5dF4D"
                 android:textColor="#000000"/>
 
             <TextView
