--- app/src/main/res/layout/activity_taking_photo.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    style="@style/loginParent"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Activities.TakingPhotoActivity">
-
-
-    <ImageView
-        android:id="@+id/iv_photo"
-        android:layout_width="395dp"
-        android:layout_height="544dp"
-        app:layout_constraintBottom_toTopOf="@+id/btn_barcode"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:src="@drawable/icon_logo" />
-
-    <Button
-        android:id="@+id/btn_barcode"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="70dp"
-        android:layout_marginEnd="70dp"
-        android:layout_marginBottom="150dp"
-        android:background="@drawable/bin_darkblue"
-        android:text="@string/cameraButton"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/btn_selfAddBook"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="70dp"
-        android:layout_marginEnd="70dp"
-        android:layout_marginBottom="50dp"
-        android:text="직접 추가"
-        android:background="@drawable/bin_darkblue"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btn_barcode" />
-</androidx.constraintlayout.widget.ConstraintLayout>
