+++ app/src/main/res/layout/layout_home.xml
@@ -15,7 +15,7 @@
 
     <!-- 버튼 디자인 수정 필요 -> 별도의 팝업이 나타나지 않는 디자인? -->
     <Button
-        android:id="@+id/btn_book"
+        android:id="@+id/btn_AddBook"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="right|top"
