+++ app/src/main/res/menu/menu_register_book.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/scanBarcode"
+        android:id="@+id/bookFind_barcode"
         android:title="카메라로 바코드 스캔" />
     <item
-        android:id="@+id/searchBookName"
+        android:id="@+id/bookFind_ISBN"
         android:title="ISBN 혹은 책 이름으로 검색" />
     <item
-        android:id="@+id/inputdirectly"
+        android:id="@+id/bookFind_direct"
         android:title="수동으로 정보 입력" />
 </menu>
