+++ app/src/main/res/values/styles.xml
@@ -9,7 +9,7 @@
 <!--        <item name="colorAccent">@color/colorAccent</item> &lt;!&ndash; No Title Bar&ndash;&gt;-->
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item> <!-- white status bar -->
-        <item name="android:statusBarColor">@color/colorPrimary</item>
+        <item name="android:windowFullscreen">true</item>
         <item name="android:windowDrawsSystemBarBackgrounds">true</item> <!-- background -->
         <item name="android:windowBackground">@drawable/splash_background</item>
     </style>
